// src/modules/proveedores/types/Proveedor.ts
export interface Proveedor {
  idProveedor?: string
  descProveedor: string
  razonSocial: string
  cifProveedor: string
  direccion: string
  codPostal: string
  poblacion: string
  provincia: string
  pais: string
  telefono: string
  fax: string
  eMail: string
  web: string
  telefono2: string
  movil: string
  formaPago: string
  ccProveedor: string
  dtoComercial: number
  observaciones: string
}